feat: 🍞 coverted docs folder into a single file
