+++ src/store/docs.js
@@ -1,11 +1,145 @@
 import router from '@/router';
 
-import DocsServices from '../../services';
-import habitatLocal from '../habitatLocal';
-import { types } from './types';
+import DocsServices from '../services';
+import habitatLocal from './habitatLocal';
 
-import { formatDocs, makeDoc } from './helpers';
-import { state } from './state';
+import store from '.';
+import { ipcRenderer } from 'electron';
+
+export const types = {
+  SET_INIT_PROJECT: 'SET_INIT_PROJECT',
+  SET_APP_PATH: 'SET_APP_PATH',
+  SET_CWD: 'SET_CWD',
+  SET_ENTRY_FILE: 'SET_ENTRY_FILE',
+  SET_DOCS_FOLDER: 'SET_DOCS_FOLDER',
+  SET_VALID_TITLE: 'SET_VALID_TITLE',
+  LOAD_DOCS: 'LOAD_DOCS',
+  ADD_DOC: 'ADD_DOC',
+  REMOVE_DOC: 'REMOVE_DOC',
+  SET_CURRENT_DOC: 'SET_CURRENT_DOC',
+  SET_TO_SAVED: 'SET_TO_SAVED',
+  SET_TO_UNSAVED: 'SET_TO_UNSAVED',
+  UPDATE_DOC_CONTENT: 'UPDATE_DOC_CONTENT',
+  SET_GUIDES: 'SET_GUIDES'
+};
+
+export const state = {
+  appPath: '',
+  cwd: '',
+  docsFolder: '',
+  entryFile: '',
+  // FIXME: Set to docsList
+  allDocs: [],
+  currentDoc: { saved: false },
+  // Register is the project is being created
+  initProject: {
+    type: undefined,
+    on: false,
+    path: ''
+  },
+  validTitle: true
+};
+
+export const mutations = {
+  /**
+   * defines if project is being initialized
+   * and in which way is being initialized
+   * @param {Object} options specifies the type of init
+   */
+  [types.SET_INIT_PROJECT](state, options) {
+    // state = {}
+    state.initProject = options;
+    console.log('SET_INIT_PROJECT options: ' + JSON.stringify(options));
+  },
+
+  [types.SET_APP_PATH](state, appPath) {
+    state.appPath = appPath;
+  },
+
+  [types.SET_CWD](state, cwd) {
+    state.cwd = cwd;
+  },
+
+  [types.SET_ENTRY_FILE](state, entryFile) {
+    state.entryFile = entryFile;
+  },
+
+  [types.SET_DOCS_FOLDER](state, docsFolder) {
+    state.docsFolder = docsFolder;
+  },
+
+  [types.SET_VALID_TITLE](state, isValid) {
+    state.validTitle = isValid;
+  },
+
+  [types.LOAD_DOCS](state, allDocs) {
+    state.allDocs = allDocs;
+    if (allDocs) {
+      state.currentDoc = allDocs[0];
+    } else {
+      state.currentDoc = undefined;
+    }
+  },
+
+  [types.ADD_DOC](state, doc) {
+    state.allDocs.push(doc);
+  },
+
+  [types.REMOVE_DOC](state, docId) {
+    const index = state.allDocs.findIndex((el) => el.id === docId);
+    state.allDocs.splice(index, 1);
+  },
+
+  [types.SET_CURRENT_DOC](state, doc) {
+    state.currentDoc = doc;
+  },
+  // FIXME: unify this mutation into SET_SAVED
+  [types.SET_TO_SAVED](state, docId) {
+    const doc = state.allDocs.find((el) => el.id === docId);
+    doc.saved = true;
+  },
+
+  [types.SET_TO_UNSAVED](state) {
+    state.currentDoc.saved = false;
+  },
+
+  [types.UPDATE_DOC_CONTENT](state, editedDoc) {
+    const newDoc = state.allDocs.find((doc) => doc.id == editedDoc.id);
+    newDoc.content = editedDoc.content;
+    newDoc.title = editedDoc.title;
+  },
+
+  [types.SET_GUIDES](state, isActive) {
+    state.guidesIsActive = isActive;
+  }
+};
+
+export const getters = {
+  docIsSaved: (state) => {
+    // console.log("Getter for isSaved " + JSON.stringify(state.currentDoc))
+    console.log('Getter for isSaved  ' + state.currentDoc.saved);
+    return state.currentDoc.saved;
+  },
+  currentDocId: (state) => {
+    return state.currentDoc.id;
+  },
+
+  hasUnsavedFiles: (state) => {
+    const allDocs = state.allDocs;
+    if (allDocs) {
+      return allDocs.filter((doc) => !doc.saved).length;
+    }
+  },
+
+  guidesIsActive: (state) => {
+    if (state.appPath === state.cwd) return true;
+    else return false;
+  },
+
+  getDocsAmount: (state) => {
+    return state.allDocs.length;
+  }
+};
 
 export const actions = {
   openFolder({ commit }) {
@@ -228,3 +362,97 @@ export const actions = {
     return DocsServices.getCWD();
   }
 };
+
+/**
+ * HELPER FUNCTIONS FOR DOCS STATE STORE
+ *
+ *
+ * Before committing the data object to the vuex it needs to be formatted
+ * The formatting includes adding an id, processing the title and
+ * adding properties such as saved.
+ * @param {Object} response the API response data object
+ * @param {Object} gqlAction this is the mutation object that wraps the data
+ */
+export function formatDocs(response, gqlAction) {
+  // console.log('formatDocs:response: ' + response.data[gqlAction])
+  let idCount = 0;
+  const allDocsData = response.data[gqlAction].allDocsData;
+  if (allDocsData) {
+    allDocsData.map((doc) => {
+      // create id
+      idCount += 1;
+      doc.id = idCount;
+
+      // Step 1: extract h1 only
+      let regex = /<[^>].+?>(.*?)<\/.+?>/m;
+      if (doc.content.match(regex)) {
+        doc.title = doc.content.match(regex)[0];
+      } else {
+        doc.title = doc.content;
+      }
+
+      // Step 2: get first block only text inside h1 tags
+      regex = /(<([^>]+)>)/gi;
+      doc.title = doc.title.replace(regex, '').trim();
+      doc.saved = true;
+      if (doc.id == 1) {
+        // Make a more unique identifier for the first document to avoid conflict with guides
+        doc.id = parseInt('' + doc.id + Math.floor(Math.random() * 1000 + 1));
+      }
+      doc.isWritten = true;
+    });
+    return allDocsData;
+  }
+}
+
+/**
+ * This function checks before a new doc object before being committed
+ * If it exists, then it appends the copy string and also creates the files
+ * accordingly with the same names.
+ * @param {Object} state to check if the new doc exists already
+ */
+export function makeDoc(state) {
+  const newId = state.allDocs.length + 1;
+  const doc = {
+    id: newId,
+    title: 'Untitled',
+    content: '',
+    saved: false
+  };
+
+  if (doc.fileName == state.entryFile) {
+    doc.fileName = state.entryFile;
+  } else {
+    // Make sure that there are no duplicate titles
+    for (var i = 0; i < state.allDocs.length; i++) {
+      if (state.allDocs[i].title == doc.title) {
+        doc.title = doc.title + ' copy';
+        doc.content = doc.title;
+      }
+    }
+    doc.content = `<h1>${doc.title}</h1>`;
+    doc['fileName'] = `${doc.title.split(' ').join('-')}.html`; // FIXME: check for duplicates
+  }
+  doc.isWritten = false;
+  return doc;
+}
+
+/**
+ * TODO: This doesnt work, try it with the plugin approach bellow....
+ */
+ipcRenderer.on('checkUnsavedDocs', () => {
+  console.log('Getting value from vuex getter to the main process');
+  let response = store.getters.hasUnsavedFiles > 0;
+  console.log('Response coming from vuex: ' + response);
+  ipcRenderer.send('hasUnsavedFiles', response);
+});
+
+// Must declare event because on render you receive an event and the data.
+// Here I should dispatch an action or maybe I should import the ipcRenderer in a component
+ipcRenderer.on('passAppPath', async (event, path) => {
+  console.log('Path coming from background process: ' + path);
+  await store.commit('SET_APP_PATH', path);
+  await store.commit('SET_CWD', path);
+  await store.dispatch('loadProject');
+  console.log(store.state.docs.cwd);
+});
