--- src/store/docs/types.js
@@ -1,16 +1,0 @@
-export const types = {
-  SET_INIT_PROJECT: 'SET_INIT_PROJECT',
-  SET_APP_PATH: 'SET_APP_PATH',
-  SET_CWD: 'SET_CWD',
-  SET_ENTRY_FILE: 'SET_ENTRY_FILE',
-  SET_DOCS_FOLDER: 'SET_DOCS_FOLDER',
-  SET_VALID_TITLE: 'SET_VALID_TITLE',
-  LOAD_DOCS: 'LOAD_DOCS',
-  ADD_DOC: 'ADD_DOC',
-  REMOVE_DOC: 'REMOVE_DOC',
-  SET_CURRENT_DOC: 'SET_CURRENT_DOC',
-  SET_TO_SAVED: 'SET_TO_SAVED',
-  SET_TO_UNSAVED: 'SET_TO_UNSAVED',
-  UPDATE_DOC_CONTENT: 'UPDATE_DOC_CONTENT',
-  SET_GUIDES: 'SET_GUIDES'
-};
