+++ test-project/.hardocs/hardocs.json
@@ -1,5 +1,5 @@
 {
-  "id": "3037473b-482c-4b1c-9028-e8714e68e5c0",
+  "id": "48c9c027-3bf5-4215-a618-329a433459ed",
   "docsDir": "docs",
   "entryFile": "index.html",
   "name": "test-project",
